+++ packages/apollo-server-hapi/src/index.ts
@@ -13,4 +13,8 @@ export {
 } from 'apollo-server-core';
 
 // ApolloServer integration.
-export { ApolloServer, ServerRegistration } from './ApolloServer';
+export {
+  ApolloServer,
+  ServerRegistration,
+  ApolloServerPluginStopHapiServer,
+} from './ApolloServer';